Move code for field properties to new file
