+++ src/app/qgsfieldsproperties.h
@@ -1,0 +1,104 @@
+/***************************************************************************
+    qgsfieldsproperties.h
+    ---------------------
+    begin                : September 2012
+    copyright            : (C) 2012 by Matthias Kuhn
+    email                : matthias dot kuhn at gmx dot ch
+ ***************************************************************************
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 2 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *                                                                         *
+ ***************************************************************************/
+
+#ifndef QGSFIELDSPROPERTIES_H
+#define QGSFIELDSPROPERTIES_H
+
+#include <QWidget>
+#include <QPushButton>
+
+#include "qgsvectorlayer.h"
+#include "ui_qgsfieldspropertiesbase.h"
+
+class QgsFieldsProperties : public QWidget, private Ui_QgsFieldsPropertiesBase
+{
+  Q_OBJECT
+
+  public:
+    QgsFieldsProperties( QgsVectorLayer *layer, QWidget* parent = 0 );
+
+    /**Adds an attribute to the table (but does not commit it yet)
+    @param field the field to add
+    @return false in case of a name conflict, true in case of success */
+    bool addAttribute( const QgsField &field );
+
+    /**Deletes an attribute (but does not commit it)
+      @param name attribute name
+      @return false in case of a non-existing attribute.*/
+    bool deleteAttribute( int attr );
+
+    void reset();
+    void apply();
+
+    void updateButtons();
+    void loadRows();
+    void setRow( int row, int idx, const QgsField &field );
+
+  signals:
+    void toggleEditing( QgsMapLayer * );
+
+  public slots:
+    void on_mAddAttributeButton_clicked();
+    void on_mDeleteAttributeButton_clicked();
+    void on_tblAttributes_cellChanged( int row, int column );
+    void on_mCalculateFieldButton_clicked();
+    void addAttribute();
+    void deleteAttribute();
+
+    void attributeAdded( int idx );
+    void attributeDeleted( int idx );
+    void attributeTypeDialog();
+
+
+
+  protected:
+    QgsVectorLayer *mLayer;
+
+    /** toggle editing of layer */
+    void toggleEditing();
+
+    /** editing of layer was toggled */
+    void editingToggled();
+
+    QMap<int, QgsVectorLayer::ValueRelationData> mValueRelationData;
+    QMap<int, QMap<QString, QVariant> > mValueMaps;
+    QMap<int, QgsVectorLayer::RangeData> mRanges;
+    QMap<int, QPair<QString, QString> > mCheckedStates;
+    QMap<int, QgsVectorLayer::EditType> mEditTypeMap;
+    QMap<int, QPushButton*> mButtonMap;
+
+    enum attrColumns
+    {
+      attrIdCol = 0,
+      attrNameCol,
+      attrTypeCol,
+      attrLengthCol,
+      attrPrecCol,
+      attrCommentCol,
+      attrEditTypeCol,
+      attrAliasCol,
+      attrWMSCol,
+      attrWFSCol,
+      attrColCount,
+    };
+
+    static QMap< QgsVectorLayer::EditType, QString > editTypeMap;
+    static void setupEditTypes();
+    static QString editTypeButtonText( QgsVectorLayer::EditType type );
+    static QgsVectorLayer::EditType editTypeFromButtonText( QString text );
+
+};
+
+#endif // QGSFIELDSPROPERTIES_H
