+++ src/app/qgsvectorlayerproperties.cpp
@@ -30,6 +30,7 @@
 #include "qgsdiagramproperties.h"
 #include "qgsdiagramrendererv2.h"
 #include "qgsfieldcalculator.h"
+#include "qgsfieldsproperties.h"
 #include "qgsgraduatedsymboldialog.h"
 #include "qgslabeldialog.h"
 #include "qgslabelinggui.h"
@@ -77,31 +78,14 @@ QgsVectorLayerProperties::QgsVectorLayerProperties(
     , mRendererDialog( 0 )
 {
   setupUi( this );
-  setupEditTypes();
 
   connect( buttonBox, SIGNAL( accepted() ), this, SLOT( accept() ) );
   connect( buttonBox, SIGNAL( rejected() ), this, SLOT( reject() ) );
   connect( buttonBox->button( QDialogButtonBox::Apply ), SIGNAL( clicked() ), this, SLOT( apply() ) );
   connect( this, SIGNAL( accepted() ), this, SLOT( apply() ) );
-  connect( mAddAttributeButton, SIGNAL( clicked() ), this, SLOT( addAttribute() ) );
-  connect( mDeleteAttributeButton, SIGNAL( clicked() ), this, SLOT( deleteAttribute() ) );
-
-  connect( mToggleEditingButton, SIGNAL( clicked() ), this, SLOT( toggleEditing() ) );
-  connect( this, SIGNAL( toggleEditing( QgsMapLayer* ) ),
-           QgisApp::instance(), SLOT( toggleEditing( QgsMapLayer* ) ) );
-
-  connect( layer, SIGNAL( editingStarted() ), this, SLOT( editingToggled() ) );
-  connect( layer, SIGNAL( editingStopped() ), this, SLOT( editingToggled() ) );
-  connect( layer, SIGNAL( attributeAdded( int ) ), this, SLOT( attributeAdded( int ) ) );
-  connect( layer, SIGNAL( attributeDeleted( int ) ), this, SLOT( attributeDeleted( int ) ) );
 
   connect( insertFieldButton, SIGNAL( clicked() ), this, SLOT( insertField() ) );
   connect( insertExpressionButton, SIGNAL( clicked() ), this, SLOT( insertExpression() ) );
-
-  mAddAttributeButton->setIcon( QgsApplication::getThemeIcon( "/mActionNewAttribute.png" ) );
-  mDeleteAttributeButton->setIcon( QgsApplication::getThemeIcon( "/mActionDeleteAttribute.png" ) );
-  mToggleEditingButton->setIcon( QgsApplication::getThemeIcon( "/mActionToggleEditing.png" ) );
-  mCalculateFieldButton->setIcon( QgsApplication::getThemeIcon( "/mActionCalculateField.png" ) );
 
   connect( btnUseNewSymbology, SIGNAL( clicked() ), this, SLOT( useNewSymbology() ) );
 
@@ -145,6 +129,10 @@ QgsVectorLayerProperties::QgsVectorLayerProperties(
   mSaveAsMenu->addAction( tr( "SLD File" ) );
   QObject::connect( mSaveAsMenu, SIGNAL( triggered( QAction * ) ), this, SLOT( saveStyleAsMenuTriggered( QAction * ) ) );
 
+  mFieldsPropertiesDialog = new QgsFieldsProperties( layer, mFieldsFrame );
+  mFieldsFrame->setLayout( new QVBoxLayout( mFieldsFrame ) );
+  mFieldsFrame->layout()->addWidget( mFieldsPropertiesDialog );
+
   reset();
 
   if ( layer->dataProvider() )//enable spatial index button group if supported by provider
@@ -173,8 +161,6 @@ QgsVectorLayerProperties::QgsVectorLayerProperties(
       grpProviderOptions->hide();
     }
   }
-
-  updateButtons();
 
   leSpatialRefSys->setText( layer->crs().authid() + " - " + layer->crs().description() );
   leSpatialRefSys->setCursorPosition( 0 );
@@ -237,134 +223,6 @@ QgsVectorLayerProperties::~QgsVectorLayerProperties()
   settings.setValue( "/Windows/VectorLayerProperties/row", tabWidget->currentIndex() );
 }
 
-void QgsVectorLayerProperties::loadRows()
-{
-  QObject::disconnect( tblAttributes, SIGNAL( cellChanged( int, int ) ), this, SLOT( on_tblAttributes_cellChanged( int, int ) ) );
-  const QgsFieldMap &fields = layer->pendingFields();
-
-  tblAttributes->clear();
-
-  tblAttributes->setColumnCount( attrColCount );
-  tblAttributes->setRowCount( fields.size() );
-  tblAttributes->setHorizontalHeaderItem( attrIdCol, new QTableWidgetItem( tr( "Id" ) ) );
-  tblAttributes->setHorizontalHeaderItem( attrNameCol, new QTableWidgetItem( tr( "Name" ) ) );
-  tblAttributes->setHorizontalHeaderItem( attrTypeCol, new QTableWidgetItem( tr( "Type" ) ) );
-  tblAttributes->setHorizontalHeaderItem( attrLengthCol, new QTableWidgetItem( tr( "Length" ) ) );
-  tblAttributes->setHorizontalHeaderItem( attrPrecCol, new QTableWidgetItem( tr( "Precision" ) ) );
-  tblAttributes->setHorizontalHeaderItem( attrCommentCol, new QTableWidgetItem( tr( "Comment" ) ) );
-  tblAttributes->setHorizontalHeaderItem( attrEditTypeCol, new QTableWidgetItem( tr( "Edit widget" ) ) );
-  tblAttributes->setHorizontalHeaderItem( attrWMSCol, new QTableWidgetItem( "WMS" ) );
-  tblAttributes->setHorizontalHeaderItem( attrWFSCol, new QTableWidgetItem( "WFS" ) );
-  tblAttributes->setHorizontalHeaderItem( attrAliasCol, new QTableWidgetItem( tr( "Alias" ) ) );
-
-  tblAttributes->horizontalHeader()->setResizeMode( 1, QHeaderView::Stretch );
-  tblAttributes->horizontalHeader()->setResizeMode( 7, QHeaderView::Stretch );
-  tblAttributes->setSelectionBehavior( QAbstractItemView::SelectRows );
-  tblAttributes->setSelectionMode( QAbstractItemView::ExtendedSelection );
-  tblAttributes->verticalHeader()->hide();
-
-  int row = 0;
-  for ( QgsFieldMap::const_iterator it = fields.begin(); it != fields.end(); it++, row++ )
-    setRow( row, it.key(), it.value() );
-
-  tblAttributes->resizeColumnsToContents();
-  QObject::connect( tblAttributes, SIGNAL( cellChanged( int, int ) ), this, SLOT( on_tblAttributes_cellChanged( int, int ) ) );
-}
-
-void QgsVectorLayerProperties::setRow( int row, int idx, const QgsField &field )
-{
-  tblAttributes->setItem( row, attrIdCol, new QTableWidgetItem( QString::number( idx ) ) );
-  tblAttributes->setItem( row, attrNameCol, new QTableWidgetItem( field.name() ) );
-  tblAttributes->setItem( row, attrTypeCol, new QTableWidgetItem( field.typeName() ) );
-  tblAttributes->setItem( row, attrLengthCol, new QTableWidgetItem( QString::number( field.length() ) ) );
-  tblAttributes->setItem( row, attrPrecCol, new QTableWidgetItem( QString::number( field.precision() ) ) );
-  tblAttributes->setItem( row, attrCommentCol, new QTableWidgetItem( field.comment() ) );
-
-  for ( int i = 0; i < attrEditTypeCol; i++ )
-    tblAttributes->item( row, i )->setFlags( tblAttributes->item( row, i )->flags() & ~Qt::ItemIsEditable );
-
-  QPushButton *pb = new QPushButton( editTypeButtonText( layer->editType( idx ) ) );
-  tblAttributes->setCellWidget( row, attrEditTypeCol, pb );
-  connect( pb, SIGNAL( pressed() ), this, SLOT( attributeTypeDialog( ) ) );
-  mButtonMap.insert( idx, pb );
-
-  //set the alias for the attribute
-  tblAttributes->setItem( row, attrAliasCol, new QTableWidgetItem( layer->attributeAlias( idx ) ) );
-
-  //published WMS/WFS attributes
-  QTableWidgetItem* wmsAttrItem = new QTableWidgetItem();
-  wmsAttrItem->setCheckState( layer->excludeAttributesWMS().contains( field.name() ) ? Qt::Unchecked : Qt::Checked );
-  wmsAttrItem->setFlags( Qt::ItemIsSelectable | Qt::ItemIsEnabled | Qt::ItemIsUserCheckable );
-  tblAttributes->setItem( row, attrWMSCol, wmsAttrItem );
-  QTableWidgetItem* wfsAttrItem = new QTableWidgetItem();
-  wfsAttrItem->setCheckState( layer->excludeAttributesWFS().contains( field.name() ) ? Qt::Unchecked : Qt::Checked );
-  wfsAttrItem->setFlags( Qt::ItemIsSelectable | Qt::ItemIsEnabled | Qt::ItemIsUserCheckable );
-  tblAttributes->setItem( row, attrWFSCol, wfsAttrItem );
-}
-
-void QgsVectorLayerProperties::attributeTypeDialog( )
-{
-  QPushButton *pb = qobject_cast<QPushButton *>( sender() );
-  if ( !pb )
-    return;
-
-  int index = mButtonMap.key( pb, -1 );
-  if ( index == -1 )
-    return;
-
-  QgsAttributeTypeDialog attributeTypeDialog( layer );
-
-  attributeTypeDialog.setValueMap( mValueMaps.value( index, layer->valueMap( index ) ) );
-  attributeTypeDialog.setRange( mRanges.value( index, layer->range( index ) ) );
-  attributeTypeDialog.setValueRelation( mValueRelationData.value( index, layer->valueRelation( index ) ) );
-
-  QPair<QString, QString> checkStates = mCheckedStates.value( index, layer->checkedState( index ) );
-  attributeTypeDialog.setCheckedState( checkStates.first, checkStates.second );
-
-  attributeTypeDialog.setIndex( index, mEditTypeMap.value( index, layer->editType( index ) ) );
-
-  if ( !attributeTypeDialog.exec() )
-    return;
-
-  QgsVectorLayer::EditType editType = attributeTypeDialog.editType();
-
-  mEditTypeMap.insert( index, editType );
-
-  QString buttonText;
-  switch ( editType )
-  {
-    case QgsVectorLayer::ValueMap:
-      mValueMaps.insert( index, attributeTypeDialog.valueMap() );
-      break;
-    case QgsVectorLayer::EditRange:
-    case QgsVectorLayer::SliderRange:
-    case QgsVectorLayer::DialRange:
-      mRanges.insert( index, attributeTypeDialog.rangeData() );
-      break;
-    case QgsVectorLayer::CheckBox:
-      mCheckedStates.insert( index, attributeTypeDialog.checkedState() );
-      break;
-    case QgsVectorLayer::ValueRelation:
-      mValueRelationData.insert( index, attributeTypeDialog.valueRelationData() );
-      break;
-    case QgsVectorLayer::LineEdit:
-    case QgsVectorLayer::TextEdit:
-    case QgsVectorLayer::UniqueValues:
-    case QgsVectorLayer::UniqueValuesEditable:
-    case QgsVectorLayer::Classification:
-    case QgsVectorLayer::FileName:
-    case QgsVectorLayer::Enumeration:
-    case QgsVectorLayer::Immutable:
-    case QgsVectorLayer::Hidden:
-    case QgsVectorLayer::Calendar:
-    case QgsVectorLayer::UuidGenerator:
-      break;
-  }
-
-  pb->setText( editTypeButtonText( editType ) );
-}
-
-
 void QgsVectorLayerProperties::toggleEditing()
 {
   emit toggleEditing( layer );
@@ -374,110 +232,6 @@ void QgsVectorLayerProperties::toggleEditing()
   if ( layer->isEditable() )
   {
     pbnQueryBuilder->setToolTip( tr( "Stop editing mode to enable this." ) );
-  }
-}
-
-void QgsVectorLayerProperties::attributeAdded( int idx )
-{
-  const QgsFieldMap &fields = layer->pendingFields();
-  int row = tblAttributes->rowCount();
-  tblAttributes->insertRow( row );
-  setRow( row, idx, fields[idx] );
-  tblAttributes->setCurrentCell( row, idx );
-}
-
-
-void QgsVectorLayerProperties::attributeDeleted( int idx )
-{
-  for ( int i = 0; i < tblAttributes->rowCount(); i++ )
-  {
-    if ( tblAttributes->item( i, 0 )->text().toInt() == idx )
-    {
-      tblAttributes->removeRow( i );
-      break;
-    }
-  }
-}
-
-void QgsVectorLayerProperties::addAttribute()
-{
-  QgsAddAttrDialog dialog( layer, this );
-  if ( dialog.exec() == QDialog::Accepted )
-  {
-    layer->beginEditCommand( "Attribute added" );
-    if ( !addAttribute( dialog.field() ) )
-    {
-      layer->destroyEditCommand();
-      QMessageBox::information( this, tr( "Name conflict" ), tr( "The attribute could not be inserted. The name already exists in the table." ) );
-    }
-    else
-    {
-      layer->endEditCommand();
-    }
-  }
-}
-
-bool QgsVectorLayerProperties::addAttribute( const QgsField &field )
-{
-  QgsDebugMsg( "inserting attribute " + field.name() + " of type " + field.typeName() );
-  layer->beginEditCommand( tr( "Added attribute" ) );
-  if ( layer->addAttribute( field ) )
-  {
-    layer->endEditCommand();
-    return true;
-  }
-  else
-  {
-    layer->destroyEditCommand();
-    return false;
-  }
-}
-
-void QgsVectorLayerProperties::deleteAttribute()
-{
-  QList<QTableWidgetItem*> items = tblAttributes->selectedItems();
-  QList<int> idxs;
-
-  for ( QList<QTableWidgetItem*>::const_iterator it = items.begin(); it != items.end(); it++ )
-  {
-    if (( *it )->column() == 0 )
-      idxs << ( *it )->text().toInt();
-  }
-  for ( QList<int>::const_iterator it = idxs.begin(); it != idxs.end(); it++ )
-  {
-    layer->beginEditCommand( tr( "Deleted attribute" ) );
-    layer->deleteAttribute( *it );
-    layer->endEditCommand();
-  }
-}
-
-void QgsVectorLayerProperties::editingToggled()
-{
-  if ( !layer->isEditable() )
-    loadRows();
-
-  updateButtons();
-}
-
-void QgsVectorLayerProperties::updateButtons()
-{
-  if ( layer->isEditable() )
-  {
-    int cap = layer->dataProvider()->capabilities();
-    mAddAttributeButton->setEnabled( cap & QgsVectorDataProvider::AddAttributes );
-    mDeleteAttributeButton->setEnabled( cap & QgsVectorDataProvider::DeleteAttributes );
-    mCalculateFieldButton->setEnabled( cap & ( QgsVectorDataProvider::ChangeAttributeValues | QgsVectorDataProvider::AddAttributes ) );
-    mToggleEditingButton->setChecked( true );
-  }
-  else
-  {
-    int cap = layer->dataProvider()->capabilities();
-    bool canChangeAttributes = cap & QgsVectorDataProvider::ChangeAttributeValues;
-    mAddAttributeButton->setEnabled( false );
-    mDeleteAttributeButton->setEnabled( false );
-    mToggleEditingButton->setChecked( false );
-    mToggleEditingButton->setEnabled( canChangeAttributes && !layer->isReadOnly() );
-    mCalculateFieldButton->setEnabled( false );
   }
 }
 
@@ -578,8 +332,6 @@ void QgsVectorLayerProperties::setDisplayField( QString name )
 //! @note in raster props, this method is called sync()
 void QgsVectorLayerProperties::reset( void )
 {
-  QObject::disconnect( tblAttributes, SIGNAL( cellChanged( int, int ) ), this, SLOT( on_tblAttributes_cellChanged( int, int ) ) );
-
   // populate the general information
   txtDisplayName->setText( layer->name() );
   pbnQueryBuilder->setWhatsThis( tr( "This button opens the query "
@@ -658,51 +410,10 @@ void QgsVectorLayerProperties::reset( void )
   //update the transparency percentage label
   sliderTransparency_valueChanged( 255 - layer->getTransparency() );
 
-  loadRows();
-  QObject::connect( tblAttributes, SIGNAL( cellChanged( int, int ) ), this, SLOT( on_tblAttributes_cellChanged( int, int ) ) );
   QObject::connect( labelCheckBox, SIGNAL( clicked( bool ) ), this, SLOT( enableLabelOptions( bool ) ) );
 } // reset()
 
 
-//
-// methods reimplemented from qt designer base class
-//
-
-QMap< QgsVectorLayer::EditType, QString > QgsVectorLayerProperties::editTypeMap;
-
-void QgsVectorLayerProperties::setupEditTypes()
-{
-  if ( !editTypeMap.isEmpty() )
-    return;
-
-  editTypeMap.insert( QgsVectorLayer::LineEdit, tr( "Line edit" ) );
-  editTypeMap.insert( QgsVectorLayer::UniqueValues, tr( "Unique values" ) );
-  editTypeMap.insert( QgsVectorLayer::UniqueValuesEditable, tr( "Unique values editable" ) );
-  editTypeMap.insert( QgsVectorLayer::Classification, tr( "Classification" ) );
-  editTypeMap.insert( QgsVectorLayer::ValueMap, tr( "Value map" ) );
-  editTypeMap.insert( QgsVectorLayer::EditRange, tr( "Edit range" ) );
-  editTypeMap.insert( QgsVectorLayer::SliderRange, tr( "Slider range" ) );
-  editTypeMap.insert( QgsVectorLayer::DialRange, tr( "Dial range" ) );
-  editTypeMap.insert( QgsVectorLayer::FileName, tr( "File name" ) );
-  editTypeMap.insert( QgsVectorLayer::Enumeration, tr( "Enumeration" ) );
-  editTypeMap.insert( QgsVectorLayer::Immutable, tr( "Immutable" ) );
-  editTypeMap.insert( QgsVectorLayer::Hidden, tr( "Hidden" ) );
-  editTypeMap.insert( QgsVectorLayer::CheckBox, tr( "Checkbox" ) );
-  editTypeMap.insert( QgsVectorLayer::TextEdit, tr( "Text edit" ) );
-  editTypeMap.insert( QgsVectorLayer::Calendar, tr( "Calendar" ) );
-  editTypeMap.insert( QgsVectorLayer::ValueRelation, tr( "Value relation" ) );
-  editTypeMap.insert( QgsVectorLayer::UuidGenerator, tr( "UUID generator" ) );
-}
-
-QString QgsVectorLayerProperties::editTypeButtonText( QgsVectorLayer::EditType type )
-{
-  return editTypeMap[ type ];
-}
-
-QgsVectorLayer::EditType QgsVectorLayerProperties::editTypeFromButtonText( QString text )
-{
-  return editTypeMap.key( text );
-}
 
 void QgsVectorLayerProperties::apply()
 {
@@ -758,77 +469,8 @@ void QgsVectorLayerProperties::apply()
 
   QSet<QString> excludeAttributesWMS, excludeAttributesWFS;
 
-  for ( int i = 0; i < tblAttributes->rowCount(); i++ )
-  {
-    int idx = tblAttributes->item( i, attrIdCol )->text().toInt();
-
-    QPushButton *pb = qobject_cast<QPushButton *>( tblAttributes->cellWidget( i, attrEditTypeCol ) );
-    if ( !pb )
-      continue;
-
-    QgsVectorLayer::EditType editType = editTypeFromButtonText( pb->text() );
-    layer->setEditType( idx, editType );
-
-    switch ( editType )
-    {
-      case QgsVectorLayer::ValueMap:
-        if ( mValueMaps.contains( idx ) )
-        {
-          QMap<QString, QVariant> &map = layer->valueMap( idx );
-          map.clear();
-          map = mValueMaps[idx];
-        }
-        break;
-
-      case QgsVectorLayer::EditRange:
-      case QgsVectorLayer::SliderRange:
-      case QgsVectorLayer::DialRange:
-        if ( mRanges.contains( idx ) )
-        {
-          layer->range( idx ) = mRanges[idx];
-        }
-        break;
-
-      case QgsVectorLayer::CheckBox:
-        if ( mCheckedStates.contains( idx ) )
-        {
-          layer->setCheckedState( idx, mCheckedStates[idx].first, mCheckedStates[idx].second );
-        }
-        break;
-
-      case QgsVectorLayer::ValueRelation:
-        if ( mValueRelationData.contains( idx ) )
-        {
-          layer->valueRelation( idx ) = mValueRelationData[idx];
-        }
-        break;
-
-      case QgsVectorLayer::LineEdit:
-      case QgsVectorLayer::UniqueValues:
-      case QgsVectorLayer::UniqueValuesEditable:
-      case QgsVectorLayer::Classification:
-      case QgsVectorLayer::FileName:
-      case QgsVectorLayer::Enumeration:
-      case QgsVectorLayer::Immutable:
-      case QgsVectorLayer::Hidden:
-      case QgsVectorLayer::TextEdit:
-      case QgsVectorLayer::Calendar:
-      case QgsVectorLayer::UuidGenerator:
-        break;
-    }
-
-    if ( tblAttributes->item( i, attrWMSCol )->checkState() == Qt::Unchecked )
-    {
-      excludeAttributesWMS.insert( tblAttributes->item( i, attrNameCol )->text() );
-    }
-    if ( tblAttributes->item( i, attrWFSCol )->checkState() == Qt::Unchecked )
-    {
-      excludeAttributesWFS.insert( tblAttributes->item( i, attrNameCol )->text() );
-    }
-  }
-
-  layer->setExcludeAttributesWMS( excludeAttributesWMS );
-  layer->setExcludeAttributesWFS( excludeAttributesWFS );
+  // Apply fields settings
+  mFieldsPropertiesDialog->apply();
 
   if ( layer->isUsingRendererV2() )
   {
@@ -1117,38 +759,6 @@ void QgsVectorLayerProperties::saveStyleAs( StyleType styleType )
   myQSettings.setValue( "style/lastStyleDir", myPath );
 }
 
-void QgsVectorLayerProperties::on_tblAttributes_cellChanged( int row, int column )
-{
-  if ( column == attrAliasCol && layer ) //only consider attribute aliases in this function
-  {
-    int idx = tblAttributes->item( row, attrIdCol )->text().toInt();
-
-    const QgsFieldMap &fields = layer->pendingFields();
-
-    if ( !fields.contains( idx ) )
-    {
-      return; // index must be wrong
-    }
-
-    QTableWidgetItem *aliasItem = tblAttributes->item( row, column );
-    if ( aliasItem )
-    {
-      layer->addAttributeAlias( idx, aliasItem->text() );
-    }
-  }
-}
-
-void QgsVectorLayerProperties::on_mCalculateFieldButton_clicked()
-{
-  if ( !layer )
-  {
-    return;
-  }
-
-  QgsFieldCalculator calc( layer );
-  calc.exec();
-}
-
 void QgsVectorLayerProperties::on_pbnSelectEditForm_clicked()
 {
   QSettings myQSettings;
@@ -1226,7 +836,6 @@ void QgsVectorLayerProperties::on_mButtonAddJoin_clicked()
       }
 
       layer->addJoin( info );
-      loadRows(); //update attribute tab
       addJoinToTreeWidget( info );
       pbnQueryBuilder->setEnabled( layer && layer->dataProvider() && layer->dataProvider()->supportsSubsetString() &&
                                    !layer->isEditable() && layer->vectorJoins().size() < 1 );
@@ -1265,7 +874,6 @@ void QgsVectorLayerProperties::on_mButtonRemoveJoin_clicked()
   }
 
   layer->removeJoin( currentJoinItem->data( 0, Qt::UserRole ).toString() );
-  loadRows();
   mJoinTreeWidget->takeTopLevelItem( mJoinTreeWidget->indexOfTopLevelItem( currentJoinItem ) );
   pbnQueryBuilder->setEnabled( layer && layer->dataProvider() && layer->dataProvider()->supportsSubsetString() &&
                                !layer->isEditable() && layer->vectorJoins().size() < 1 );
