+++ src/app/qgsvectorlayerproperties.h
@@ -39,6 +39,7 @@ class QgsVectorLayer;
 class QgsVectorOverlayPlugin;
 class QgsLabelingGui;
 class QgsDiagramProperties;
+class QgsFieldsProperties;
 
 class QgsVectorLayerProperties : public QDialog, private Ui::QgsVectorLayerPropertiesBase
 {
@@ -79,8 +80,6 @@ class QgsVectorLayerProperties : public QDialog, private Ui::QgsVectorLayerPrope
 
     void insertExpression();
 
-    void attributeTypeDialog();
-
     void alterLayerDialog( const QString& string );
 
     /** Reset to original (vector layer) values */
@@ -98,12 +97,6 @@ class QgsVectorLayerProperties : public QDialog, private Ui::QgsVectorLayerPrope
     /** Called when apply button is pressed or dialog is accepted */
     void apply();
 
-    /** toggle editing of layer */
-    void toggleEditing();
-
-    /** editing of layer was toggled */
-    void editingToggled();
-
     //
     //methods reimplemented from qt designer base class
     //
@@ -115,19 +108,12 @@ class QgsVectorLayerProperties : public QDialog, private Ui::QgsVectorLayerPrope
     void on_pbnSaveDefaultStyle_clicked();
     void on_pbnLoadStyle_clicked();
     void on_pbnSaveStyleAs_clicked();
-    void on_tblAttributes_cellChanged( int row, int column );
-    void on_mCalculateFieldButton_clicked();
     void on_pbnSelectEditForm_clicked();
     void on_tabWidget_currentChanged( int idx );
     void on_buttonBox_helpRequested() { QgsContextHelp::run( metaObject()->className() ); }
     void on_pbnUpdateExtents_clicked();
 
     void enableLabelOptions( bool theFlag );
-    void addAttribute();
-    void deleteAttribute();
-
-    void attributeAdded( int idx );
-    void attributeDeleted( int idx );
 
     void useNewSymbology();
     void setUsingNewSymbology( bool useNewSymbology );
@@ -143,6 +129,8 @@ class QgsVectorLayerProperties : public QDialog, private Ui::QgsVectorLayerPrope
     void toggleEditing( QgsMapLayer * );
 
   private slots:
+    /** toggle editing of layer */
+    void toggleEditing();
 
     /** save the style based on selected format from the menu */
     void saveStyleAsMenuTriggered( QAction * );
@@ -152,21 +140,6 @@ class QgsVectorLayerProperties : public QDialog, private Ui::QgsVectorLayerPrope
     void saveStyleAs( StyleType styleType );
 
     void updateSymbologyPage();
-
-    enum attrColumns
-    {
-      attrIdCol = 0,
-      attrNameCol,
-      attrTypeCol,
-      attrLengthCol,
-      attrPrecCol,
-      attrCommentCol,
-      attrEditTypeCol,
-      attrAliasCol,
-      attrWMSCol,
-      attrWFSCol,
-      attrColCount,
-    };
 
     QgsVectorLayer *layer;
 
@@ -186,18 +159,11 @@ class QgsVectorLayerProperties : public QDialog, private Ui::QgsVectorLayerPrope
     QgsAttributeActionDialog* actionDialog;
     /**Diagram dialog. If apply is pressed, options are applied to vector's diagrams*/
     QgsDiagramProperties* diagramPropertiesDialog;
+    /**Fields dialog. If apply is pressed, options are applied to vector's diagrams*/
+    QgsFieldsProperties* mFieldsPropertiesDialog;
+
 
     QList<QgsApplyDialog*> mOverlayDialogs;
-    QMap<int, QPushButton*> mButtonMap;
-    QMap<int, QgsVectorLayer::EditType> mEditTypeMap;
-    QMap<int, QMap<QString, QVariant> > mValueMaps;
-    QMap<int, QgsVectorLayer::RangeData> mRanges;
-    QMap<int, QgsVectorLayer::ValueRelationData> mValueRelationData;
-    QMap<int, QPair<QString, QString> > mCheckedStates;
-
-    void updateButtons();
-    void loadRows();
-    void setRow( int row, int idx, const QgsField &field );
 
     void initDiagramTab();
 
@@ -209,11 +175,6 @@ class QgsVectorLayerProperties : public QDialog, private Ui::QgsVectorLayerPrope
 
     /**Adds a new join to mJoinTreeWidget*/
     void addJoinToTreeWidget( const QgsVectorJoinInfo& join );
-
-    static QMap< QgsVectorLayer::EditType, QString > editTypeMap;
-    static void setupEditTypes();
-    static QString editTypeButtonText( QgsVectorLayer::EditType type );
-    static QgsVectorLayer::EditType editTypeFromButtonText( QString text );
 };
 
 inline QString QgsVectorLayerProperties::displayName()
